moved scenes into appropriate packages/modules
