+++ fxgl-cutscene/src/main/java/module-info.java
@@ -3,6 +3,9 @@
  */
 module com.almasb.fxgl.cutscene {
     requires com.almasb.fxgl.core;
+    requires com.almasb.fxgl.animation;
+    requires com.almasb.fxgl.input;
+    requires com.almasb.fxgl.scene;
 
     exports com.almasb.fxgl.cutscene;
 }
