+++ fxgl/src/main/kotlin/com/almasb/fxgl/app/Settings.kt
@@ -15,7 +15,7 @@ import com.almasb.fxgl.core.serialization.Bundle
 import com.almasb.fxgl.core.util.Optional
 import com.almasb.fxgl.core.util.Platform
 import com.almasb.fxgl.core.util.RuntimeInfo
-import com.almasb.fxgl.dsl.CutsceneService
+import com.almasb.fxgl.cutscene.CutsceneService
 import com.almasb.fxgl.notification.impl.NotificationServiceProvider
 import com.almasb.fxgl.notification.view.NotificationView
 import com.almasb.fxgl.notification.view.XboxNotificationView
